+++ src/JWT/CustomTokenGenerator.php
@@ -7,7 +7,6 @@ namespace Kreait\Firebase\JWT;
 use DateInterval;
 use Kreait\Firebase\JWT\Action\CreateCustomToken;
 use Kreait\Firebase\JWT\Action\CreateCustomToken\Handler;
-use Kreait\Firebase\JWT\Action\CreateCustomToken\WithHandlerDiscovery;
 use Kreait\Firebase\JWT\Contract\Token;
 use Kreait\Firebase\JWT\Error\CustomTokenCreationFailed;
 use Kreait\Firebase\JWT\Value\Duration;
@@ -28,7 +27,7 @@ final class CustomTokenGenerator
 
     public static function withClientEmailAndPrivateKey(string $clientEmail, string $privateKey): self
     {
-        $handler = new WithHandlerDiscovery($clientEmail, $privateKey, SystemClock::fromUTC());
+        $handler = new CreateCustomToken\WithLcobucciJWT($clientEmail, $privateKey, SystemClock::fromUTC());
 
         return new self($handler);
     }