+++ src/JWT/Keys/ExpiringKeys.php
@@ -5,15 +5,17 @@ declare(strict_types=1);
 namespace Kreait\Firebase\JWT\Keys;
 
 use DateTimeImmutable;
+use DateTimeInterface;
 use Kreait\Firebase\JWT\Contract\Expirable;
-use Kreait\Firebase\JWT\Contract\ExpirableTrait;
 use Kreait\Firebase\JWT\Contract\Keys;
-use Kreait\Firebase\JWT\Contract\KeysTrait;
 
 final class ExpiringKeys implements Expirable, Keys
 {
-    use ExpirableTrait;
-    use KeysTrait;
+    /** @var array<string, string> */
+    private $values = [];
+
+    /** @var DateTimeImmutable */
+    private $expirationTime;
 
     private function __construct()
     {
@@ -30,5 +32,31 @@ final class ExpiringKeys implements Expirable, Keys
         $keys->expirationTime = $expirationTime;
 
         return $keys;
+    }
+
+    public function withExpirationTime(DateTimeImmutable $expirationTime): Expirable
+    {
+        $expirable = clone $this;
+        $expirable->expirationTime = $expirationTime;
+
+        return $expirable;
+    }
+
+    public function isExpiredAt(DateTimeInterface $now): bool
+    {
+        return $this->expirationTime < $now;
+    }
+
+    public function expiresAt(): DateTimeImmutable
+    {
+        return $this->expirationTime;
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    public function all(): array
+    {
+        return $this->values;
     }
 }