+++ src/JWT/Keys/StaticKeys.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace Kreait\Firebase\JWT\Keys;
 
 use Kreait\Firebase\JWT\Contract\Keys;
-use Kreait\Firebase\JWT\Contract\KeysTrait;
 
 final class StaticKeys implements Keys
 {
-    use KeysTrait;
+    /** @var array<string, string> */
+    private $values = [];
 
     private function __construct()
     {
@@ -29,5 +29,13 @@ final class StaticKeys implements Keys
         $keys->values = $values;
 
         return $keys;
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    public function all(): array
+    {
+        return $this->values;
     }
 }