+++ tests/JWT/Action/FetchGooglePublicKeys/WithGuzzleTest.php
@@ -10,7 +10,7 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use Kreait\Firebase\JWT\Action\FetchGooglePublicKeys\Handler;
-use Kreait\Firebase\JWT\Action\FetchGooglePublicKeys\WithGuzzle6;
+use Kreait\Firebase\JWT\Action\FetchGooglePublicKeys\WithGuzzle;
 use Kreait\Firebase\JWT\Error\FetchingGooglePublicKeysFailed;
 use Kreait\Firebase\JWT\Keys\ExpiringKeys;
 use Kreait\Firebase\JWT\Keys\StaticKeys;
@@ -18,7 +18,7 @@ use Kreait\Firebase\JWT\Keys\StaticKeys;
 /**
  * @internal
  */
-final class WithGuzzle6Test extends TestCase
+final class WithGuzzleTest extends TestCase
 {
     /** @var MockHandler */
     private $mockHandler;
@@ -32,7 +32,7 @@ final class WithGuzzle6Test extends TestCase
 
     protected function createHandler(): Handler
     {
-        return new WithGuzzle6(new Client(['handler' => $this->mockHandler]), $this->clock);
+        return new WithGuzzle(new Client(['handler' => $this->mockHandler]), $this->clock);
     }
 
     /**